BuildImages: fix mystery whitespace in build logs
In some forms the "Building image1 image2" message is interleaved with
whitespace.  This is because the description field is meant to be a list but
this class sets it as a string.  Thanks to duck typing the list operations to
join it with " " work fine, and the message ends up spaced.  Fix this by setting
the description as a list.

Signed-off-by: Ross Burton <ross.burton@intel.com>